requirements.txt
asgiref==3.3.4
colorama==0.4.4
Django==3.2
django-phone-field==1.8.1
django-phonenumber-field==5.1.0
django-quill-editor==0.1.22
django-wysiwyg==0.8.0
phonenumbers==8.12.22
Pillow==8.2.0
pytz==2021.1
qrcode==6.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
